refactor to split components in Extractor ⭐️
this is a great example of how fine-grained modules and zustand can enhance the performace of react app
(i.e. to drasticaly reduce the chances of re-render)
even no  React.memo() is need.

useRef also takes a very important role here, sometimes a small input can cause the whole app to re-render again and again while typing,
this is definitely not acceptable for me.

in my premature thoughts, useState is not for every situation and, react forget to make an API for situations where a  "state" should be tracked, but it won't be used in rendering the UI. this is where all the <input> type of elements are in, because they dont need rerender to affect the value change. if an input element's value is not used anywhere in the UI, IMO, it should never use useState to store and change its value.
in the future, I can explore some more simpler and elegant way to write `const inputRef = useRef(null); const getInputValue = () => inputRef.current.value`, maybe create a library for it.

OK, two topics for blog is prepared.